Reduced allowed line max length to 100
As discussed in one of our NinJS guild meeting, the allowed max length
for a single line of code is now 100 characters.

See the issue here: https://github.com/samsao/guilds/issues/106

Additionally, I added some rules that were not defined in our config but
were new to the eslint-react-plugin. I tried to update the airbnb config
we use before doing so but the changes aren't published yet. (The last
npm publish was two months ago and the changes to add the rules were
commited a month ago). We need to keep an eye on the next
eslint-config-airbnb release.