Changed peerDependencies into normal dependencies
In our specific case, having normal dependencies reduce the maintenance
burden of specifying all the needed peer dependencies when adding a
specific ESLint configuration.

That's means that we simply need to depend on the right config project
(mobile, node or web) to run ESLint.

Of course, it could causes conflict problems eventually, but we are ready to
live with it.